renamed xml and json modules to xml_utils and json_utils to prevent name collisions with standard library modules
