+++ jsonize/__init__.py
@@ -1,3 +1,3 @@
-from .utils.json import JSONNodeType, JSONNode, JSONPath
-from .utils.xml import XPath, XMLNodeType, XMLNode
+from .utils.json_utils import JSONNodeType, JSONNode, JSONPath
+from .utils.xml_utils import XPath, XMLNodeType, XMLNode
 from .utils.mapping import Transformation, XMLNodeToJSONNode, xml_document_to_json
