+++ jsonize/utils/xml_utils.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from enum import Enum
 from typing import Tuple, Dict, Iterable, Union, List
-from .json import JSONPath
+from .json_utils import JSONPath
 from xml.etree import ElementTree
 import re
